fortescrow-backend/requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.5
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
cron-descriptor==1.4.0
Django==4.2.7
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==12.0.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.4
model-mommy==2.0.0
msgpack==1.0.7
packaging==23.2
paystack==1.5.0
paystackapi==2.1.0
pluggy==1.3.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycparser==2.21
pytest==7.4.3
pytest-django==4.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
